+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See Git checking messages for full history.
 ## 9.0.2 (2023/xx/xx)
 - leveled up the packaging using `hatchling`
 - used `ruff` to lint the code base (#275)
+- MSS: allow to use a custom date function for output file names (#276)
 - MSS: minor optimization when using an output file format without date (#275)
 - MSS: fixed `Pixel` model type (#274)
 - CI: automated release publishing on tag creation